+++ src/Facades/Events.lua
@@ -1,16 +1,18 @@
---[[
+--[[--
 The Events class is a layer between World of Warcraft events and events
 triggered by the Stormwind Library.
 
 When using this library in an addon, it should focus on listening to the
 library events, which are more detailed and have more mapped parameters.
+
+@classmod Core.Events
 ]]
 local Events = {}
     Events.__index = Events
     Events.__ = self
     self:addClass('Events', Events)
 
-    --[[
+    --[[--
     Events constructor.
     ]]
     function Events.__construct()
@@ -30,10 +32,12 @@ local Events = {}
         return self
     end
 
-    --[[
+    --[[--
     Creates the events frame, which will be responsible for capturing
     all World of Warcraft events and forwarding them to the library
     handlers.
+
+    @local
     ]]
     function Events:createFrame()
         self.eventsFrame = CreateFrame('Frame')
@@ -42,13 +46,17 @@ local Events = {}
         end)
     end
 
-    --[[
+    --[[--
     This is the main event handler method, which will capture all
     subscribed World of Warcraft events and forwards them to the library
     handlers that will later notify other subscribers.
 
     It's important to mention that addons shouldn't care about this
     method, which is an internal method to the Events class.
+
+    @tparam table source The Events instance, used when calling Events.handleOriginal
+    @tparam string event The World of Warcraft event to be handled
+    @param ... The parameters passed by the World of Warcraft event
     ]]
     function Events:handleOriginal(source, event, ...)
         local callback = self.originalListeners[event]
@@ -58,7 +66,7 @@ local Events = {}
         end
     end
 
-    --[[
+    --[[--
     Listens to a Stormwind Library event.
 
     This method is used by addons to listen to the library events.
@@ -73,7 +81,7 @@ local Events = {}
         table.insert(self.listeners[event], callback)
     end
 
-    --[[
+    --[[--
     Sets the Events Frame to listen to a specific event and also store the
     handler callback to be called when the event is triggered.
 
@@ -88,19 +96,22 @@ local Events = {}
         self.originalListeners[event] = callback
     end
 
-    --[[
+    --[[--
     Notifies all listeners of a specific event.
 
     This method should be called by event handlers to notify all listeners
     of a specific Stormwind Library event.
+
+    @tparam string event The Stormwind Library event to notify
+    @param ... The parameters to be passed to the event listeners
     ]]
     function Events:notify(event, ...)
-        local params = ...
+        local params = {...}
 
         local listeners = self.__.arr:get(self.listeners, event, {})
 
         self.__.arr:map(listeners, function (listener)
-            listener(params)
+            listener(self.__.arr:unpack(params))
         end)
     end
 -- end of Events