+++ src/Models/Macro.lua
@@ -1,15 +1,17 @@
---[[
+--[[--
 The macro class maps macro information and allow in game macro updates.
+
+@classmod Models.Macro
 ]]
 local Macro = {}
     Macro.__index = Macro
     Macro.__ = self
     self:addClass('Macro', Macro)
 
-    --[[
+    --[[--
     Macro constructor.
 
-    @tparam string name the macro's name
+    @tparam string name The macro's name
     ]]
     function Macro.__construct(name)
         local self = setmetatable({}, Macro)
@@ -22,16 +24,16 @@ local Macro = {}
         return self
     end
 
-    --[[
+    --[[--
     Determines whether this macro exists.
 
-    @treturn boolean
+    @treturn boolean Whether the macro exists
     ]]
     function Macro:exists()
         return GetMacroIndexByName(self.name) > 0
     end
 
-    --[[
+    --[[--
     Saves the macro, returning the macro id.
 
     If the macro, identified by its name, doesn't exist yet, it will be created.
@@ -43,7 +45,7 @@ local Macro = {}
     Future implementations may fix this issue, but as long as it uses unique
     names, this model will work as expected.
 
-    @treturn integer the macro id
+    @treturn integer The macro id
     ]]
     function Macro:save()
         if self:exists() then
@@ -53,7 +55,7 @@ local Macro = {}
         return CreateMacro(self.name, self.icon, self.body)
     end
 
-    --[[
+    --[[--
     Sets the macro body.
 
     The macro's body is the code that will be executed when the macro's
@@ -62,36 +64,36 @@ local Macro = {}
     If the value is an array, it's considered a multiline body, and lines will
     be separated by a line break.
 
-    @tparam array<string>|string value the macro's body
+    @tparam array[string]|string value The macro's body
 
-    @return self
+    @treturn Models.Macro self The current instance for method chaining
     ]]
     function Macro:setBody(value)
         self.body = self.__.arr:implode('\n', value)
         return self
     end
 
-    --[[
+    --[[--
     Sets the macro icon.
 
-    @tparam integer|string value the macro's icon texture id
+    @tparam integer|string value The macro's icon texture id
 
-    @return self
+    @treturn Models.Macro self The current instance for method chaining
     ]]
     function Macro:setIcon(value)
         self.icon = value
         return self
     end
 
-    --[[
+    --[[--
     Sets the macro name.
 
     This is the macro's identifier, which means the one World of Warcraft API
     will use when accessing the game's macro.
 
-    @tparam string value the macro's name
+    @tparam string value The macro's name
 
-    @return self
+    @treturn Models.Macro self The current instance for method chaining
     ]]
     function Macro:setName(value)
         self.name = value